Added solution of second iteration
Added three more algorithms, one with privatized histograms in shared
memory, one with multiple input chars done by single warp, and one
combining these two.

Results on lorem.txt:
privatized - 17.7
aggregated - 111 with high variability
privatized_aggregated - 0.9
- 1 private copy, 128 itemsPerThread

Multiple private histograms do not make sense, as the access to
histogram entries is random, so collisions are not that common anyway
and it just adds additional collisions when collecting the results into
the global memory.
- maybe if we merged the results locally and then pushed it to the
global memory

aggregated by itself is no improvement over atomics, as the bottleneck
there is of course access to the global memory

The combination works best, as can be seen in the resutls.
As threads do more work, the privatization becomes even better and it
just runs faster.