Added block local aggregation of results
Instead of adding all copies of the local histogram separately to the
global memory, now we sum them locally and then update the global memory
once.

This should reduce the number of global memory accesses and collisions.

It does remove the performance regression with increasing number of
local copies, keeping the performance stable, but there is not
discernable increase in performace from having multiple copies.

This means that the bottleneck is somewhere else.