Fixed privatization, fixed existing solutions
Fixed privatization, so that it now prevents bank conflicts. Now we check that the size of the histogram is not divisible or does not divide 32, so that successive copies are not aligned in the banks.

Changed the way copies are assigned to threads. As conflicts only arise in single warp, we need to split threads of a warp to work with different copies. Now we divide the threads using their thread id modulo number of copies.

Changed aggregation of results from the private copies. Instead of some halving and stuff, we just have every thread aggregate one value in the resulting histogram from all copies. This does not need any syncthreads and works nice.

Fixed naming of existing solutions to match the assignment. Also added blockSize to all solutions, as that is the only variable that makes sense for every solution.