Removed old misleading comments, moved shared mem declaration
Removed old misleading comments which were left over from previous
implementations and were no longer true.

Changed the shared memory declaration from being global variable to
local variable for each kernel using shared memory. It just seems like a
better way of doing it. Also removes an ugly reinterpret cast, as when
it was outside the functions, we could not access the RES template
parameter and so had to cast the shared memory to the correct pointer
type.