Fixed overlap algorithm
Fixed overlap algorithm as mentioned in Issue #1 . Previous
implementation did not overlap anything, as every operation was sent to
the 0 stream, in which the operations were run sequentialy by the GPU.

The new implementation adds CUDA streams into the mix. Processing of
each chunk is now basically completely independent on other chunks,
possibly running in parallel. The cudaMemcpyOperation and kernel start
for the given chunk are passed into the same CUDA stream, ensuring
proper ordering so that the input data is available to the kernel.

Added commandline parameter numStreams, controlling the number of CUDA
streams which are used to dispatch the chunk processing operations.
Streams are used in round robin fashion. This means that the chunk 0
will be dispatched to stream <streams[0]>, chunk 1 to <streams[1]>,
up to <numStreams - 1> chunk, where chunk <numStreams> will be
dispatched again into the <stream[0]>. Simply chunk modulo numStreams.

This also allows us to only allocate <numStreams> * <chunkSize> memory
for the input data on the GPU, as only <numStreams> chunks will ever run
in parallel. Other chunks will be scheduled in the same streams after
them, so the other chunks can then reuse the memory from the previous
chunks in the same stream. Also there may be situations where
<numChunks> is less than <numStreams>, so we allocate only min of these
two values number of chunks in the device memory. Because there is only
<numChunks> chunks in this situation, we can be sure that no chunks will
be scheduled to streams above <numChunks>, so noone will access the
memory which would have been allocated for the unused streams.

Added stream argument to run_atomic_shm, enabling us to start the kernel
in streams other than 0. This parameter has default value 0, to preserve
the implementation of other algorithms without code changes.

Simplified the implementation of CudaOverlapAlgorithm::run method.
Instead of having separate part of the method dealing with the last
chunk, which may be smaller than <chunkSize>, we track the number of
items left to be processed. With each chunk dispatch, we set the chunk
size to min of <chunkSize> and the <itemsToProcess>, which ensures that
the last chunk will process only the items left.

As we can see from the changes in README.md, the changes described above
lead to speedup of 16%, which is nice.